Python/mongodb/Pymongo-Part-2/requirements.txt
dns-python
pymongo